+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 
+## 0.1.2
+
+- Added const extractor
+
 ## 0.1.1
 
 - Fixed exports